+++ balancer-js/src/index.ts
@@ -3,10 +3,10 @@ export * from './pool-weighted';
 export * from './pool-utils';
 export * from './utils';
 export * from './types';
-export * from './swapsService/index';
-export * from './swapsService/types';
-export * from './swapsService/helpers';
+export * from './services/swaps/swaps.service';
+export * from './services/swaps/types';
+export * from './services/swaps/helpers';
 export * from './constants/network';
 export * from './constants/subgraph';
 export * from './sdk';
-export * from './relayerService/index';
+export * from './services/relayer/relayer.service';
